Move ROOT_VIEW to common/utils to improve waterfall
This avoids an extra request on initial load for components that use that variable
